Fix 0.0.0.0/0 route management
If subnet['gateway_ip'] is not None, Neutron prevents manual setting of
0.0.0.0/0 route to subnet host routes.

So we set gateway_ip to None during subnet creation if not internet
gateway is used in an associated route table. As a result Neutron
doesn't add its own 0.0.0.0/0 route, and we don't need to redirect it to
127.0.0.1 if default route is not set for a subnet route table. It
simplifies mapping of VPC route table to Neutron subnet host routes.

Depends-On: I5c314d513922d8a21b3e63de632bb9cd3484b076
Change-Id: Ibeeea1f5d92cc0aa08ac6c360c3e73fa67a45373