Re-adding the py35 testing job
ec2api-tempest-plugin is branchless and test all the supported
stable branches including which run on xenial(before stein).

That means ec2api-tempest-plugin code should be able to run on py35
and so does we need py35 testing job which got removed as part of
updating the Train PTI runtime update
- https://review.opendev.org/#/c/667087

This commit re-adding the py35 job.
Story: #2005924
Task: #34206

Change-Id: Iab2ff1e4037608b3afca37a2effbf50485d5491c