[ussuri][goal] Drop python 2.7 support and testing
OpenStack is dropping the py2.7 support in ussuri cycle.

ec2api-tempest-plugin is ready with python 3 and ok to drop the
python 2.7 support.

Removing the USE_PYTHON3=True from base job so that it can
keep running  on py3 for < ussuri branches.

Complete discussion & schedule can be found in
- http://lists.openstack.org/pipermail/openstack-discuss/2019-October/010142.html
- https://etherpad.openstack.org/p/drop-python2-support

Ussuri Communtiy-wide goal:
https://governance.openstack.org/tc/goals/selected/ussuri/drop-py27.html

Change-Id: I49feaa3eddf66058ebd061b76b76172cb4373c3e